safari-600.7-branch/LayoutTests/platform/ios-sim/webgl/conformance/context/context-lost-restored-expected.txt

CONSOLE MESSAGE: line 104: TypeError: null is not an object (evaluating 'extension.loseContext')
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance/context/context-lost-restored.html
PASS
